+++ datumbox-framework-core/src/test/java/com/datumbox/framework/core/machinelearning/regression/MatrixLinearRegressionTest.java
@@ -124,6 +124,7 @@ public class MatrixLinearRegressionTest extends AbstractTest {
         
         LinearRegressionMetrics vm = new Validator<>(LinearRegressionMetrics.class, configuration)
                 .validate(new KFoldSplitter(k).split(trainingData), param);
+        System.out.println(vm);
 
         double expResult = 1;
         double result = vm.getRSquare();